requirements.txt
langchain-community==0.0.3 ; python_version >= "3.10" and python_version < "4.0"
langchain-core==0.1.1 ; python_version >= "3.10" and python_version < "4.0"
langchain==0.0.350 ; python_version >= "3.10" and python_version < "4.0"
langsmith==0.0.71 ; python_version >= "3.10" and python_version < "4.0"
pytest-base-url==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytest-playwright==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.4.3 ; python_version >= "3.10" and python_version < "4.0"
openai==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
html2text==2020.1.16 ; python_version >= "3.10" and python_version < "4.0"
streamlit==1.29.0 ; python_version >= "3.10" and python_version < "4.0"
pypdf2==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
pyperclip==1.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
